agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-google
python-dotenv
pydantic-settings>=2.7.1
livekit-api
flask[async]
flask
flask-cors
uvicorn
beanie
ollama